requirements.txt
pillow==10.2.0
paddlepaddle==2.6.0
paddleocr==2.7.0.3
openpyxl==3.1.2
loguru==0.7.2